Calendar refactor
Calendar has been renamed to Month
+ Calendar component now hosts both Month & LeaveOverviewGrid
+ Calendar & LeaveOverviewGrid now rely on route params
